msm: iommu: [GCC6] Avoid -Wshift-overflow (SET_FIELD is 32-bit)

+++ arch/arm/mach-msm/include/mach/iommu_hw-v1.h
@@ -58,8 +58,8 @@
 #define SET_FIELD(addr, mask, shift, v) \
 do { \
 	int t = readl_relaxed(addr); \
-	writel_relaxed((t & ~((mask) << (shift))) + (((v) & \
-			(mask)) << (shift)), addr); \
+	writel_relaxed((t & ~(((u32) mask) << (shift))) + (((v) & \
+			((u32) mask)) << (shift)), addr); \
 } while (0)
